pyproject.toml
[project]
name = "youtube-object-detection"
version = "2.2.0"
description = "Real-time object detection for YouTube live streams with YOLOv11 and Hungarian algorithm tracking"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "opencv-python>=4.12.0.88",
    "requests>=2.32.5",
    "yt-dlp>=2025.8.22",
    "numpy>=1.23.0",
    "pillow>=7.1.2",
    "ultralytics>=8.3.0",
]

# Clean configuration without heavy PyTorch dependencies
# YOLOv11 runs in browser using TensorFlow.js instead
